Adding ICMP interface
This commit adds an interface in Network to send and receive
ICMP packets. This interface is very simple and It allows
only to read one ICMP at time. I also added an test named
ToroPing.pas which sends an ICMP packet and waits for a
response. One ICMP packets is sent every 2s.