Memory: Support regions above the QEMU hole (#519)
This commit allows to allocate memory per core by relying on multiples
memory regions. Before this commit, only the first usable region was used to
allocate memory for all the cores. However, this does not work well when
there are holes in the physical memory, e.g., in QEMU. This commit
allows to allocate memory from different regions for the same core.

Signed-off-by: Matias Ezequiel Vara Larsen <matiasevara@torokernel.io>