Networking: Reduce to one the number of copies when sending packets (#363)
This commit reduces to one the number of copies required when sending
network packets. The SysNetworkSend() function owns the Packet and it is
the Network stack that releases the whole Packet structure. This enables
the VSocket driver to reuse Packet.Data to send the packet. To do this,
the VirtIO buffer uses the pointer at Packet.Data instead of copying the
content to a new buffer. After device uses the buffer, the driver
informs the kernel that the Packet have been sent thus releasing the
Packet by the kernel.

Signed-off-by: Matias Ezequiel Vara Larsen <matiasevara@torokernel.io>